// frontend/src/api/gen/types/JoinDecisionAction.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export const joinDecisionActionEnum = {
    approve: "approve",
    reject: "reject",
    reject_and_block: "reject_and_block"
} as const;

export type JoinDecisionActionEnumKey = (typeof joinDecisionActionEnum)[keyof typeof joinDecisionActionEnum];

/**
 * JoinDecisionAction
 * @description What a leader does with a pending request.
*/
export type JoinDecisionAction = JoinDecisionActionEnumKey;